【SY01-V2.1】修复flash读写问题

+++ App_User/compass_manage.c
@@ -3,7 +3,7 @@
 extern  float Target_Azimuth; //校准后的最终角度
 
 //标定偏移值
-extern float Calibration_Offset; //校准偏移量
+extern int32_t Calibration_Offset; //校准偏移量
